+++ management/vo/label.go
@@ -6,7 +6,7 @@ import (
 )
 
 type LabelVo struct {
-	ID          uint           `json:"id"`
+	ID          uint64         `json:"id"`
 	Label       string         `json:"name"`
 	CreatedAt   time.Time      `json:"createdAt"`
 	DeletedAt   gorm.DeletedAt `json:"deletedAt"`
@@ -19,8 +19,8 @@ type LabelVo struct {
 // NodeLabelVo Node label relation
 type NodeLabelVo struct {
 	ModelVo
-	NodeId    uint
-	LabelId   uint
+	NodeId    uint64
+	LabelId   uint64
 	LabelName string
 	CreatedBy string
 	UpdatedBy string
